// frontend/src/sections/help/helpDocuments.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type HelpDocumentsVariables = Record<string, never>;

export type HelpDocumentsResult = {
  helpDocuments: ({
  __typename: "HelpDocumentConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  title: string;
  createdDatetime: string;
  files: {
  __typename: "SyncFileReferenceConnector";
  nodes: Array<{
  id: string;
  fileName: string;
  mimeType: string | null;
  totalBytes: number;
}>;
};
}>;
});
};

export const HelpDocuments = {
  query: "query helpDocuments {\n  helpDocuments(page: {first: 1000, offset: 0}) {\n    ... on HelpDocumentConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        title\n        createdDatetime\n        files {\n          ... on SyncFileReferenceConnector {\n            __typename\n            nodes {\n              id\n              fileName\n              mimeType\n              totalBytes\n            }\n          }\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<HelpDocumentsResult, HelpDocumentsVariables>;