// frontend/src/sections/help/contactForm.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../api/graphql";

export type InsertContactFormVariables = {
  storeId: string;
  input: {
    id: string;
    contactType: "FEEDBACK" | "SUPPORT";
    replyEmail: string;
    body: string;
  };
};

export type InsertContactFormResult = {
  insertContactForm: ({
  __typename: "InsertResponse";
} & {
  id: string;
});
};

export const InsertContactForm = {
  query: "mutation insertContactForm($storeId: String!, $input: InsertContactFormInput!) {\n  insertContactForm(storeId: $storeId, input: $input) {\n    __typename\n    ... on InsertResponse {\n      id\n    }\n  }\n}",
} as TypedDocument<InsertContactFormResult, InsertContactFormVariables>;